+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/domain/collections/MapEntry.kt
@@ -1,0 +1,7 @@
+package com.windea.breezeframework.core.domain.collections
+
+data class MapEntry<K, V>(
+	override val key: K,
+	override val value: V
+) : Map.Entry<K, V>
+