+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/domain/collections/MultiValueSet.kt
@@ -1,5 +1,0 @@
-package com.windea.breezeframework.core.domain.collections
-
-//TODO
-
-interface MultiValueSet<E> : Collection<E>