+++ CHANGELOG.md
@@ -174,6 +174,9 @@
 ### 1.0.7
 
 * [X] 更新版本。
+* [X] ~~breeze-dsl 提供`SteamTextDsl`。~~
+* [X] breeze-core 基本`MultiValueMap`。
+* [ ] breeze-dsl 分割为多个子模块。
 * [ ] breeze-dsl 优化：对于表示转换的元素，可以通过`"a"(...) fromTo "b"`的语法构建。
 * [ ] breeze-dsl 实现`YamlDsl`。（不推荐使用，但是作为规范）
 ***